@@ -1,24 +1,14 @@
-﻿namespace TauCode.Db.DbValueConverters
+﻿using System;
+
+namespace TauCode.Db.DbValueConverters
 {
     public class DecimalValueConverter : DbValueConverterBase
     {
         protected override object ToDbValueImpl(object value)
         {
-            if (value is decimal decimalValue)
+            if (value.GetType().IsNumericType())
             {
-                return decimalValue;
-            }
-            else if (value is double doubleValue)
-            {
-                return doubleValue;
-            }
-            else if (value is int intValue)
-            {
-                return intValue;
-            }
-            else if (value is long longValue)
-            {
-                return longValue;
+                return Convert.ToDecimal(value);
             }
 
             return null;
